requirements.txt
blinker==1.6.2
click==8.1.7
dotenv-python==0.0.1
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
Werkzeug==2.3.7
